Format Oroboros O2K Output.csv Into Lists

o2k_oxy_to_list() extracts data from a csv file which has been produced by DatLab and the Oroboros O2K. The output is a list with three elements. Element one is itself a list of two elements (left and right), which contain the raw data and events for the left and right chambers respectively. Element two is a tibble containing details of when the chambers were opened; the third and final element is a vector of unique events from within the run (excluding 'open', 'closed', and 'warnings'). The results of this function are designed so that they can be directly passed into o2k_oxy_list_to_tbl().

Usage

o2k_oxy_to_list(csv = "NULL")

Arguments

csv

A csv file exported from the Oroboros-O2K. This should have six columns ("Time s", "Event Name", "Chamber", "Event Text", "1A: O2 flux per V pmol/(s*mL)" and "1B: O2 flux per V pmol/(s*mL))" ).

Value

A list of 3 elements; a list of raw oxygen flux measurements per chamber, opening time details, and unique events.

See also

Examples

 data <- tidymito::o2k_oxy_to_list(csv = "path/to/data/file.csv")
#> Error: 'path/to/data/file.csv' does not exist in current working directory ('/private/var/folders/32/lg72_x752993pxg2fvmzrcr00000gn/T/RtmpmJFM6f/filef1141f359106/reference').